+++ js/Demo/demoTree.js
@@ -45,5 +45,8 @@ addNode("h", {
 )
 
 addLayer("tree-tab", {
-    tabFormat: [["tree", function() {return (layoutInfo.treeLayout ? layoutInfo.treeLayout : TREE_LAYERS)}]]
+    tabFormat: [["tree", function() {return (layoutInfo.treeLayout ? layoutInfo.treeLayout : TREE_LAYERS)}]],
+    previousTab: "",
+    leftTab: true,
+
 })